% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bau_mmr_all_countries_proj.R
\name{bau_mmr_all_countries_proj}
\alias{bau_mmr_all_countries_proj}
\title{Calculate BAU MMR projections for all countries for a given range}
\usage{
bau_mmr_all_countries_proj(mmr_pivotwider_tibble, arr_start_year_colnum,
  arr_end_year_colnum, mmr_start_year, mmr_end_year)
}
\arguments{
\item{mmr_pivotwider_tibble}{A tibble that includes all countries' observed MMR in the pivotwider format}

\item{arr_start_year_colnum}{The column number of the year to start the BAU ARR calculation with}

\item{arr_end_year_colnum}{The column number of the to end the BAU ARR calculation with}

\item{mmr_start_year}{A year to start BAU MMR projections from}

\item{mmr_end_year}{A year to end BAU MMR projections on}
}
\value{
The BAU MMR projections for all countries for the specified year range
}
\description{
Calculate BAU MMR projections for all countries for a given range
}
