% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bau_mmr_regional_projection_summaries.R
\name{bau_mmr_regional_projection_summaries}
\alias{bau_mmr_regional_projection_summaries}
\title{Get BAU MMR Projections for a Specified Period by Region and Globally}
\usage{
bau_mmr_regional_projection_summaries(mmr_pivotwider_tibble,
  countries_regions_tibble, birth_projections_tibble,
  arr_start_year_colnum, arr_end_year_colnum, proj_start_year,
  proj_end_year)
}
\arguments{
\item{mmr_pivotwider_tibble}{A tibble that contains all countries' MMR in the pivotwider format}

\item{countries_regions_tibble}{A tibble that lists all countries and the region they belong to}

\item{birth_projections_tibble}{A tibble with the births by country for 2030}

\item{arr_start_year_colnum}{The column number of the year to start the BAU ARR calculation with}

\item{arr_end_year_colnum}{The column number of the to end the BAU ARR calculation with}

\item{proj_start_year}{A year to begin the summary projections from}

\item{proj_end_year}{A year to end the summary projections with}
}
\value{
A tibble summarizing regional and global BAU MMR projections for a specified time period
}
\description{
Get BAU MMR Projections for a Specified Period by Region and Globally
}
