% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mmr_sdg_proj.R
\name{get_mmr_sdg_proj}
\alias{get_mmr_sdg_proj}
\title{Calculate the target year's MMR projections, Adjusted for No Country's MMR Above 140, based on a chosen ARR}
\usage{
get_mmr_sdg_proj(mmr2015_country, arr, nproject)
}
\arguments{
\item{mmr2015_country}{A vector with MMR for 2015 for all countries}

\item{arr}{A number (the ARR, intended to be the calculated SDG ARR)}

\item{nproject}{Number of years to project, number of years since the base year, 2015}
}
\value{
The MMR projections based on a chosen ARR, adjusted so no country has an MMR over 140
}
\description{
Calculate the target year's MMR projections, Adjusted for No Country's MMR Above 140, based on a chosen ARR
}
