% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmr_proj_all_regions.R
\name{mmr_proj_regional}
\alias{mmr_proj_regional}
\title{Calculate MMR projections by region}
\usage{
mmr_proj_regional(mmr_pivotwider_tibble, mmr_proj_tibble,
  countries_regions_tibble, births_iso, sdg1_region, proj_start_year,
  proj_end_year)
}
\arguments{
\item{mmr_pivotwider_tibble}{A tibble that includes all countries' observed MMR in the pivotwider format}

\item{mmr_proj_tibble}{A tibble of the projected MMRs for all countries during the desired year range}

\item{countries_regions_tibble}{A tibble that lists all countries and the region they belong to}

\item{births_iso}{A tibble with the births by country ISO code for the desired year}

\item{sdg1_region}{The exact name of the desired region (must be in quotes)}

\item{proj_start_year}{A year to begin the summary projections from}

\item{proj_end_year}{A year to end the summary projections with}
}
\value{
The MMR projections by region for the specified year range
}
\description{
Calculate MMR projections by region
}
