% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmr_proj_single_country.R
\name{mmr_proj_single_country}
\alias{mmr_proj_single_country}
\title{Calculate the MMR projections for a specified country}
\usage{
mmr_proj_single_country(mmr_pivotwider_tibble, arr_tibble, arr_col_num,
  iso_code, mmr_start_year, mmr_end_year)
}
\arguments{
\item{mmr_pivotwider_tibble}{A tibble that includes all countries' observed MMR in the pivotwider format}

\item{arr_tibble}{A tibble that includes the desired ARR for every country along w/ each country's ISO code}

\item{arr_col_num}{The column number of the column of arr_tibble that lists each country's ARR}

\item{iso_code}{The ISO Code of a chosen country}

\item{mmr_start_year}{A year to start MMR projections from}

\item{mmr_end_year}{A year to end MMR projections on}
}
\value{
The MMR projections for one country for the specified year range
}
\description{
Calculate the MMR projections for a specified country
}
