% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdg_mmr_all_countries_proj.R
\name{sdg_mmr_allcountries_proj}
\alias{sdg_mmr_allcountries_proj}
\title{Calculate the BAU MMR projections for all countries for a given period}
\usage{
sdg_mmr_allcountries_proj(mmr_pivotwider_tibble, bau_sdg_arr_tibble,
  mmr_start_year, mmr_end_year)
}
\arguments{
\item{mmr_pivotwider_tibble}{A tibble that includes all countries' observed MMR in the pivotwider format}

\item{bau_sdg_arr_tibble}{A tibble that includes all of the BAU and SDG ARRs}

\item{mmr_start_year}{A year to start BAU MMR projections from}

\item{mmr_end_year}{A year to end BAU MMR projections on}
}
\value{
The SDG MMR projections for all countries for the specified year range
}
\description{
Calculate the BAU MMR projections for all countries for a given period
}
