% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_arr_sdg_target.R
\name{get_arr_sdg_target}
\alias{get_arr_sdg_target}
\title{Calculate the ARR SDG Target (to later get country-specific arrs)}
\usage{
get_arr_sdg_target(mmr2015_country, births_tibble, nproject)
}
\arguments{
\item{mmr2015_country}{A vector with MMR for 2015 for all countries}

\item{births_tibble}{A tibble with the births by country for 2030}

\item{nproject}{The number of years to project, number of years since the base year, 2015}
}
\value{
Optimizes the SDG ARR to help with the calculation of SDG ARRs for each country
}
\description{
Calculate the ARR SDG Target (to later get country-specific arrs)
}
